pyproject.toml
[tool.poetry]
name = "histoprep"
version = "2.0.0"
description = "Read and process histological slide images with python!"
authors = ["jopo666 <jopo@birdlover.com>"]
readme = "README.md"

[tool.poetry.scripts]
HistoPrep-cut = 'histoprep.cli._run:cut'

[tool.poetry.dependencies]
aicspylibczi = "^3"
click = "^8"
matplotlib = "*"
numpy = "*"
opencv-python-headless = "^4.7.0.68"
openslide-python = "^1.2"
pillow = "^9"
python = "^3.9"
tqdm = "*"
mpire = "^2.6.0"
polars = "^0.16.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "*"
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
